Add Mode types
