+++ lib/mode.ts
@@ -1,5 +1,5 @@
 import TfLAPI from './tfl';
-import { GetArrivalPredictionsAllStops, GetActiveServiceTypes } from './interfaces/mode';
+import * as IMode from './interfaces/IMode';
 
 export default class Mode extends TfLAPI {
     constructor(config: string) {
@@ -9,7 +9,7 @@ export default class Mode extends TfLAPI {
     /**
      * Returns the service type active for a mode. Currently only supports tube
      */
-    getActiveServiceTypes(): Promise<GetActiveServiceTypes.Root[]> {
+    getActiveServiceTypes(): Promise<IMode.GetActiveServiceTypes.Root> {
         return this.sendRequest(`/Mode/ActiveServiceTypes`, {}, 'GET');
     }
 
@@ -18,7 +18,7 @@ export default class Mode extends TfLAPI {
      * @param mode A mode name e.g. tube, dlr
      * @param count Number of arrivals to return for each stop, -1 to return all available (default).
      */
-    getArrivalPredictionsAllStops(mode: string, count: number = -1): Promise<GetArrivalPredictionsAllStops.Root[]> {
+    getArrivalPredictionsAllStops(mode: string, count: number = -1): Promise<IMode.GetArrivalPredictionsAllStops.Root> {
         return this.sendRequest(`/Mode/${mode}/Arrivals`, { count }, 'GET');
     }
 }
